refactor: implement OID4VC client structure and integrate with existing modules
