+++ crates/cloud-wallet-openid4vc/src/core/mod.rs
@@ -1,1 +1,2 @@
 pub mod claim_path_pointer;
+pub mod client;
