+++ crates/cloud-wallet-openid4vc/src/oid4vci/client/mod.rs
@@ -10,16 +10,11 @@ pub use signer::{
     Algorithm, Claims as ProofClaims, CryptoSigner, Header as ProofHeader, ProofSigner,
 };
 
-use std::sync::Arc;
-use std::time::Duration;
-
 use futures::stream::{FuturesUnordered, StreamExt};
 use reqwest::header::CONTENT_TYPE;
-use reqwest_middleware::{ClientBuilder, ClientWithMiddleware};
-use reqwest_retry::policies::ExponentialBackoff;
-use reqwest_retry::{Jitter, RetryTransientMiddleware};
 use url::Url;
 
+use crate::core::client::OidClient;
 use crate::oid4vci::authorization::{
     AuthorizationDetails, AuthorizationRequest, AuthorizationResponse, CodeChallengeMethod,
     PushedAuthorizationRequest, PushedAuthorizationResponse,
@@ -48,8 +43,6 @@ use crate::utils::pkce::{derive_pkce_challenge, generate_pkce_verifier};
 
 type Result<T> = std::result::Result<T, ClientError>;
 
-const HTTP_MAX_RETRIES: u32 = 3;
-const DEFAULT_HTTP_TIMEOUT_SECS: u64 = 10;
 const OAUTH_RESPONSE_TYPE: &str = "code";
 const FORM_ENCODED_HEADER: &str = "application/x-www-form-urlencoded";
 const JSON_HEADER: &str = "application/json";
@@ -136,87 +129,6 @@ impl AuthorizationCallback {
     }
 }
 
-/// Configuration for the OID4VCI client.
-///
-/// # Security
-///
-/// - Outbound requests are HTTPS-only and not configurable.
-/// - For testing purposes, hostname validation can be temporarily disabled via
-///   `accept_untrusted_hosts(true)`.
-#[derive(Debug, Clone)]
-pub struct Config {
-    /// The wallet's OAuth 2.0 `client_id` as registered at the issuer AS.
-    pub client_id: String,
-    /// The redirect URI registered with the issuer AS.
-    pub redirect_uri: Url,
-    /// Total timeout for each request.
-    pub timeout: Duration,
-    /// Optional user-agent value to send with every request.
-    pub user_agent: Option<String>,
-    /// Accept untrusted hosts (testing only).
-    ///
-    /// **WARNING**: This bypasses TLS certificate validation and HTTPS enforcement.
-    /// Only use in test environments with mock servers. Never enable in production.
-    pub accept_untrusted_hosts: bool,
-    /// Use system proxy configuration discovered by the HTTP client.
-    pub use_system_proxy: bool,
-}
-
-impl Config {
-    /// Creates a new configuration with the given client ID and redirect URI.
-    ///
-    /// Defaults:
-    /// - timeout: 10 seconds
-    /// - user_agent: None
-    /// - accept_untrusted_hosts: false
-    pub fn new(client_id: impl Into<String>, redirect_uri: Url) -> Self {
-        Self {
-            client_id: client_id.into(),
-            redirect_uri,
-            timeout: Duration::from_secs(DEFAULT_HTTP_TIMEOUT_SECS),
-            user_agent: None,
-            accept_untrusted_hosts: false,
-            use_system_proxy: true,
-        }
-    }
-
-    /// Sets the total request timeout.
-    ///
-    /// Defaults to 10 seconds.
-    pub fn timeout(self, timeout: Duration) -> Self {
-        Self { timeout, ..self }
-    }
-
-    /// Sets a custom user-agent header value.
-    pub fn user_agent(self, user_agent: impl Into<String>) -> Self {
-        Self {
-            user_agent: Some(user_agent.into()),
-            ..self
-        }
-    }
-
-    /// Enables or disables accepting untrusted hosts.
-    ///
-    /// This should only be enabled in test environments.
-    pub fn accept_untrusted_hosts(self, accept_untrusted_hosts: bool) -> Self {
-        Self {
-            accept_untrusted_hosts,
-            ..self
-        }
-    }
-
-    /// Enables or disables system proxy discovery.
-    ///
-    /// Disabling this is useful in tests and restricted runtime environments
-    /// where reading host networking configuration may fail during client setup.
-    pub fn use_system_proxy(self, use_system_proxy: bool) -> Self {
-        Self {
-            use_system_proxy,
-            ..self
-        }
-    }
-}
-
 /// High-level OID4VCI orchestration client.
 ///
 /// The client covers:
@@ -230,52 +142,15 @@ impl Config {
 /// - Notification (optional, best-effort)
 #[derive(Debug, Clone)]
 pub struct Oid4vciClient {
-    config: Arc<Config>,
-    http_client: ClientWithMiddleware,
+    inner_client: OidClient,
 }
 
 impl Oid4vciClient {
-    /// Creates a new client with custom HTTP options for the internal request client.
-    pub fn new(config: Config) -> Result<Self> {
-        let retry_policy = ExponentialBackoff::builder()
-            .jitter(Jitter::Bounded)
-            .build_with_max_retries(HTTP_MAX_RETRIES);
-
-        let mut inner_client_builder = reqwest::Client::builder()
-            .timeout(config.timeout)
-            .tls_backend_rustls()
-            .https_only(true)
-            .tls_danger_accept_invalid_hostnames(config.accept_untrusted_hosts);
-
-        if config.accept_untrusted_hosts {
-            inner_client_builder = inner_client_builder.tls_danger_accept_invalid_certs(true);
+    /// Create a new OID4VCI client from an OID client.
+    pub fn new(client: OidClient) -> Self {
+        Self {
+            inner_client: client,
         }
-
-        if !config.use_system_proxy {
-            inner_client_builder = inner_client_builder.no_proxy();
-        }
-
-        if let Some(ref user_agent) = config.user_agent {
-            inner_client_builder = inner_client_builder.user_agent(user_agent);
-        }
-
-        let inner_client = inner_client_builder.build().map_err(|e| {
-            ClientError::configuration(format!("failed to build HTTP client: {e:?}"))
-        })?;
-
-        let http_client = ClientBuilder::new(inner_client)
-            .with(RetryTransientMiddleware::new_with_policy(retry_policy))
-            .build();
-
-        Ok(Self {
-            config: Arc::new(config),
-            http_client,
-        })
-    }
-
-    /// Returns the underlying HTTP client.
-    pub fn http_client(&self) -> &ClientWithMiddleware {
-        &self.http_client
     }
 
     /// Resolve a raw credential offer string into a typed `CredentialOffer`.
@@ -290,7 +165,7 @@ impl Oid4vciClient {
             CredentialOfferSource::ByValue(offer) => Ok(offer),
             CredentialOfferSource::ByReference(url) => {
                 // For by-reference offers, we need to fetch them
-                resolve_by_reference(&url, &self.http_client)
+                resolve_by_reference(&url, &self.inner_client.http_client())
                     .await
                     .map_err(ClientError::from)
             }
@@ -310,7 +185,8 @@ impl Oid4vciClient {
         let url = well_known_issuer_metadata_url(credential_issuer);
 
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .get(url)
             .header(CONTENT_TYPE, JSON_HEADER)
             .send()
@@ -350,7 +226,8 @@ impl Oid4vciClient {
         let rfc8414_url = well_known_oauth_as_url(&as_url);
         // Try the primary endpoint
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .get(rfc8414_url)
             .send()
             .await
@@ -374,7 +251,8 @@ impl Oid4vciClient {
         // Fallback: openid-configuration (RFC 8414 §5 / OIDC Discovery)
         let oidc_url = well_known_oidc_configuration_url(&as_url);
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .get(oidc_url)
             .send()
             .await
@@ -463,8 +341,8 @@ impl Oid4vciClient {
         // Build the authorization request
         let authz_request = AuthorizationRequest {
             response_type: OAUTH_RESPONSE_TYPE.into(),
-            client_id: self.config.client_id.clone(),
-            redirect_uri: Some(self.config.redirect_uri.clone()),
+            client_id: self.inner_client.config().client_id.clone(),
+            redirect_uri: Some(self.inner_client.config().redirect_uri.clone()),
             state: Some(state.into()),
             scope: None,
             resource: None,
@@ -481,7 +359,11 @@ impl Oid4vciClient {
         {
             // PAR is advertised, use it.
             let par_response = self.send_par(context, &authz_request).await?;
-            build_par_redirect_url(authz_endpoint, &self.config.client_id, par_response)?
+            build_par_redirect_url(
+                authz_endpoint,
+                &self.inner_client.config().client_id,
+                par_response,
+            )?
         } else {
             build_plain_authz_url(authz_endpoint, authz_request)?
         };
@@ -550,8 +432,8 @@ impl Oid4vciClient {
 
         let request = TokenRequest::AuthorizationCode(AuthorizationCodeRequest {
             code: code.into(),
-            redirect_uri: Some(self.config.redirect_uri.clone()),
-            client_id: self.config.client_id.clone(),
+            redirect_uri: Some(self.inner_client.config().redirect_uri.clone()),
+            client_id: self.inner_client.config().client_id.clone(),
             code_verifier: Some(pkce_verifier.into()),
             authorization_details: Some(authz_details),
         });
@@ -582,7 +464,7 @@ impl Oid4vciClient {
         let client_id = if context.as_metadata.allows_anonymous_pre_authorized_grant() {
             None
         } else {
-            Some(self.config.client_id.clone())
+            Some(self.inner_client.config().client_id.clone())
         };
 
         // Include authorization_details so the AS can return credential_identifiers
@@ -606,7 +488,8 @@ impl Oid4vciClient {
         }
 
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .post(nonce_endpoint.as_str())
             .send()
             .await
@@ -711,7 +594,8 @@ impl Oid4vciClient {
         };
 
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .post(endpoint.as_str())
             .bearer_auth(access_token)
             .json(&request)
@@ -753,7 +637,8 @@ impl Oid4vciClient {
         req: &NotificationRequest,
     ) -> Result<()> {
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .post(notification_endpoint.as_str())
             .bearer_auth(access_token)
             .json(&req)
@@ -843,7 +728,8 @@ impl Oid4vciClient {
             .map_err(|e| ClientError::internal(format!("Failed to serialize PAR request: {e}")))?;
 
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .post(par_endpoint.as_str())
             .header(CONTENT_TYPE, FORM_ENCODED_HEADER)
             .body(body)
@@ -882,7 +768,8 @@ impl Oid4vciClient {
         })?;
 
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .post(token_endpoint.as_str())
             .header(CONTENT_TYPE, FORM_ENCODED_HEADER)
             .body(body)
@@ -925,7 +812,7 @@ impl Oid4vciClient {
         let client_id = if is_anonymous {
             None
         } else {
-            Some(self.config.client_id.to_owned())
+            Some(self.inner_client.config().client_id.to_owned())
         };
 
         let claims = ProofClaims {
@@ -946,7 +833,8 @@ impl Oid4vciClient {
         request: &CredentialRequest,
     ) -> Result<CredentialResponse> {
         let response = self
-            .http_client
+            .inner_client
+            .http_client()
             .post(context.issuer_metadata.credential_endpoint.as_str())
             .bearer_auth(access_token)
             .json(request)
